# openapi.yaml
openapi: 3.0.0
info:
  title: Fusion Stock Analyst API
  description: AI-enhanced stock analysis platform with real-time signals and KPIs
  version: 1.7.4
  contact:
    name: Fusion Stock Analyst
    email: support@fusionanalyst.com
servers:
  - url: http://0.0.0.0:5000
    description: Development server
paths:
...
  /api/papertrade/orders:
    get:
      tags:
        - Paper Trade
      summary: Get paper trade order history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Order history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaperTradeOrder'
  /api/papertrade/execute:
    post:
      tags:
        - Paper Trade
      summary: Execute paper trade at live market price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - symbol
                - side
                - quantity
              properties:
                symbol:
                  type: string
                  example: "RELIANCE"
                side:
                  type: string
                  enum: [BUY, SELL]
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Trade executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  order:
                    $ref: '#/components/schemas/PaperTradeOrder'
                  live_price:
                    type: number
  /api/papertrade/close:
    post:
      tags:
        - Paper Trade
      summary: Close paper trade position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - symbol
              properties:
                symbol:
                  type: string
                  example: "RELIANCE"
      responses:
        '200':
          description: Position closed successfully
...
components:
  schemas:
...
    PaperTradeOrder:
      type: object
      properties:
        order_id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        quantity:
          type: integer
        order_type:
          type: string
        exec_price:
          type: number
        exec_value:
          type: number
        timestamp:
          type: string
          format: date-time
        status:
          type: string
...